backend/HERA/requirements.txt
python-telegram-bot>=20.0
paho-mqtt>=2.0
ollama>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
paho-mqtt==1.6.1
